movie-rag/requirements.txt
pandas==2.1.4              
numpy==1.26.4              
sentence-transformers==5.0.0
faiss-cpu==1.11.0.post1    
